misc/cgo: fix issue 9400 test on 386
issue9400_linux.go did not build on 386 because it used a constant
that was larger than a 32-bit int in a ... argument.  Fix this by
casting the constant to uint64 (to match how the constant is being
used).

Reviewed-on: https://go-review.googlesource.com/2060

+++ misc/cgo/test/issue9400_linux.go
@@ -52,7 +52,7 @@ func test9400(t *testing.T) {
 	// Check test pattern
 	for i := range big {
 		if big[i] != pattern {
-			t.Fatalf("entry %d of test pattern is wrong; %#x != %#x", i, big[i], pattern)
+			t.Fatalf("entry %d of test pattern is wrong; %#x != %#x", i, big[i], uint64(pattern))
 		}
 	}
 }
